moved EditableCategory to QueryResult namespace
